+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - `Innmind\Cli\Console::error()` now return an `Innmind\Immutable\Attempt`
 - `Innmind\Cli\Environment::output()` now return an `Innmind\Immutable\Attempt`
 - `Innmind\Cli\Environment::error()` now return an `Innmind\Immutable\Attempt`
+- `Innmind\Cli\Question\Question` now return the read data in an `Innmind\Immutable\Attempt`
+- `Innmind\Cli\Question\ChoiceQuestion` now return the read data in an `Innmind\Immutable\Attempt`
 
 ### Fixed